XuanQi: fixed no argument configs/ directory scan bug
It appears when running XuanQi without any argument it fails
to bring up the help menu when it supposed to. The bug was
caused by the recent 'configs/' directory scan requirement
and missed out the no argument situation. Hence, let's fix it.

This patch fixes no argument configs/ directory scan bug in
XuanQi.

Co-authored-by: Shuralyov, Jean <jean.shuralyov@proton.me>
Co-authored-by: Galyna, Cory <cory.galyna@gmail.com>
Co-authored-by: 湘渝萱 <alice.xiang.yu.xuan@gmail.com>
Co-authored-by: (Holloway) Chew, Kean Ho <me@hollowaykeanho.com>
Reviewed-by: (Holloway) Chew, Kean Ho <me@hollowaykeanho.com>
Signed-off-by: (Holloway) Chew, Kean Ho <me@hollowaykeanho.com>